CSS: use flexbox, move html+body rules to index.html
Instead of `float`ing, it's better to use the new flexbox. Instead of `position:fixed`, better set the heights to `100vh` (i.e. 100% of the viewport height) and set `overflow-y:auto` so we get scrollbars.
Move CSS for `html` and `body` to the index.html file because it doesn't really apply to the component.