Modularize DocGen comp. into several sub-components
The new components (one each for the three parts of the viewer: TOC, links section and processes section) may not have much custom behaviour, but this allows the re-use of individual sub-components in other projects like the openEO Hub. The ProcessPanel component is now exported too, so that other tools can display the spec of single individual processes.