Limit CSS rules to the #docgen container
If styling is applied to `html` it is highly likely that the entire styling of a site that reuses the component will be ruined.
I also prepended `#docgen` to all rules that didn't have it yet so we avoid conflicts (`csList` may not be a very common class name, but `required` certainly is).
I didn't prepend it to something like `#docgen-toc-and-links-container` because that should be unique enough :D