Remove doubled name, add hub-specific metadata
Maybe there's a reason why the process's name is printed twice, but if so, I don't get it. The name is already printed *right above*, so I don't see a point in doing it again. Also, it doesn't work nice with my hub layout. So I don't think it's a loss if it is removed.

In the hub, I need to display two more metadata properties: Which backend a rendered process belongs to (users can search for processes across all backends) and at what time the Hub crawled the data from the backend (i.e. how up-to-date is the Hub's cache). I added `v-if` directives to the elements, so this change will not be noticed in the standalone viewer.