Add "show more/less" button aka collapse feature
I display lots of ProcessPanels in a long list without a TOC, so I want to be able to collapse the details when not necessary. Wrapping everything under the summary (or after my last commit: the backend name) in a div and putting a `v-show` on it plus a "show more/less" button for UI unsurprisingly does the trick.

I added a prop `initiallyCollapsed`. If it's set to true the details are collapsed and there's a button to expand them. If it's set to false the details are all shown but nevertheless there's the button to hide them. **If it's not set at all or undefined or null, all the details are shown and there is no button at all, i.e. everything works as if this commit never had come into existence.**